chore: Break down and organize __init__.py
